A door is a metaphor. Knock, bust, or find a key.
My new roommate (name still unknown) who moved in yesterday, upon coming
home tonight locked the door behind him on the way in. This speaks
volumes of his personality and/or habits, given that we live on the 6th
floor of a private apartment building with a keyfob to get into the
front door, manned (or womanned in this case) by someone at the front
desk.

His return overlapped with my ~2 minute journey to take my laundry out
of the dryer. Of course I don't have my keys... So now I'm locked out.
This isn't the first time fate has spun this mishap either, however last
time it was my roommate Ashley leaving, and I managed to get to her
before she left the building.

The *fucked up* part is what happened next. I proceed to knock on the
front door many times, hoping he had in fact returned home. I even went
so far as to go up to his window (a floor above) and knock on the glass
directly. Somehow it didn't occur to him that he should go check the
door however, despite admitting he had heard the knocks and taps. He's
not a mean person I'm sure, and he seems genuinely sorry for my
inconvenience. Something about strangers, or the possibility of
strangers is too much to be bothered with for some people I suppose.